field: the lead's and the party's ability are read whole
EncounterGenState keeps the lead Pokemon's ability in a u8, and
GetEggCyclesToSubtract reads each party member's into a u8 local. Both come
from GetMonData(MON_DATA_ABILITY), which returns all nine bits, so an ability
past 255 was compared by its low byte against the retail field abilities.
Farigiraf, reachable from the wild Girafarig that learns Twin Beam, is the
live case: Cud Chew (291) doubled the encounter rate as Illuminate (35) and
Armor Tail (296) halved egg cycles as Magma Armor (40). Palafin's 278 read as
Intimidate, Toxic Chain (305) as Flame Body, and so on down the list.

The reference does not hook either routine, so its game compares the same
low byte; the fix follows the abilities' own meaning instead. The state is
private to encounter_check.c, lives only on the stack and no assembly reads
it, so its field becomes a u16; the egg sentinel NUM_ABILITIES still fits.
The egg-cycle local becomes a u16 too.

Engine layer: the abilities past 255 are the engine's range.

tests/newgold/test_field_ability_width.py runs EncounterGenState_Init,
ApplyAbilityEffectToEncounterRate and GetEggCyclesToSubtract natively for
abilities 0..511 and checks that only the exact retail ids change the rate or
the cycles. Reverting either file makes it fail. Both ROMs build; the suite
passes except the ledger counter.

+++ src/field/encounter_check.c
@@ -56,7 +56,7 @@ typedef struct EncounterGenState {
     BOOL isSweetScent;
     u8 level;
     u8 isEgg;
-    u8 ability;
+    u16 ability;
     u8 unk_0F;
     u8 unk_10;
     u8 isSinjohMap;
